Update PATH when adding commands in a .mk include.
This will allow recipes to work with just the basename of commands.

Of course, a .mk file must be brought in with `include` for the PATH change
to take effect, making these PATH updates opt-in.

Subprocesses inherit PATH; run a configured shell with a recipe:

    shell:
        @exec /bin/bash

A solution to generalize which shell (bash, zsh, ...) would have to query
the OS (with `getent` on Linux and `dscl` on Mac OS X) because Makefile
uses $SHELL; if it didn't, $SHELL would provide the answer on which
shell to start interactively.

Instead of generalizing the solution in Makefile.d, add a recipe to the
project which starts the shell of choice. In order to have a configured
interactive shell, use a custom shell recipe, as in the example above.